pyproject.toml
[project]
name = "fastapi-general-framework"
version = "0.1.0"
description = "compose general components for fastapi framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aliyun-python-sdk-core>=2.15.1",
    "esdk-obs-python>=3.25.3",
    "fastapi==0.115.13",
    "httpx>=0.28.1",
    "oss2>=2.19.1",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "streamlit>=1.50.0",
    "uvicorn==0.34.3",
    "sqlalchemy~=2.0.34",
    "sqlmodel~=0.0.27",
    "ffmpeg-python==0.2.0",
    "pydub",
    "celery[redis]",
    "loguru",
    "pymediainfo",
    "diskcache",
    "yarl>=1.23.0",
    "av>=17.0.0",
    "func-to-web>=0.9.13",
    "pymysql",
    "opensearch-py",
    "aio-pika",
    "aiomysql",
    "python-dotenv",
    "scenedetect[opencv]",
    "opencv-python"
]
